Revert "Merge pull request #26 from Unice-SIL/feature/bo-admin-dupliquer-syllabus"
This reverts commit 9fcb95c490d57fbeb2f2e3566e05a2f5dd56e52f, reversing
changes made to 585f6dd1caa27204bac1832b17be0a61c2ef4c08.

+++ src/AppBundle/Query/Course/EditEquipmentsCourseInfoQuery.php
@@ -2,18 +2,18 @@
 
 namespace AppBundle\Query\Course;
 
-use AppBundle\Entity\CourseInfo;
-use AppBundle\Entity\CourseResourceEquipment;
+use AppBundle\Command\Course\EditEquipmentsCourseInfoCommand;
+use AppBundle\Exception\CourseInfoNotFoundException;
+use AppBundle\Query\QueryInterface;
 use AppBundle\Repository\CourseInfoRepositoryInterface;
 use AppBundle\Repository\CourseResourceEquipmentRepositoryInterface;
-use Ramsey\Uuid\Uuid;
 use Symfony\Component\Security\Core\Security;
 
 /**
  * Class EditEquipmentsCourseInfoQuery
  * @package AppBundle\Query\Course
  */
-class EditEquipmentsCourseInfoQuery
+class EditEquipmentsCourseInfoQuery implements QueryInterface
 {
 
     /**
@@ -25,6 +25,11 @@ class EditEquipmentsCourseInfoQuery
      * @var CourseResourceEquipmentRepositoryInterface
      */
     private $courseResourceEquipmentRepository;
+
+    /**
+     * @var EditEquipmentsCourseInfoCommand
+     */
+    private $editEquipmentsCourseInfoCommand;
 
     /**
      * @var Security
@@ -49,37 +54,45 @@ class EditEquipmentsCourseInfoQuery
     }
 
     /**
-     * @param CourseInfo $courseInfo
-     * @param CourseInfo $originalCourseInfo
+     * @param EditEquipmentsCourseInfoCommand $editEquipmentsCourseInfoCommand
+     * @return EditEquipmentsCourseInfoQuery
+     */
+    public function setEditEquipmentsCourseInfoCommand(EditEquipmentsCourseInfoCommand $editEquipmentsCourseInfoCommand): EditEquipmentsCourseInfoQuery
+    {
+        $this->editEquipmentsCourseInfoCommand = $editEquipmentsCourseInfoCommand;
+        return $this;
+    }
+
+    /**
+     * @throws CourseInfoNotFoundException
      * @throws \Exception
      */
-    public function execute(CourseInfo $courseInfo, CourseInfo $originalCourseInfo): void
+    public function execute(): void
     {
+        try {
+            // Find CourseInfo
+            $courseInfo = $this->courseInfoRepository->find($this->editEquipmentsCourseInfoCommand->getId());
+        }catch (\Exception $e){
+            throw $e;
+        }
+        if(is_null($courseInfo)){
+            throw new CourseInfoNotFoundException(sprintf('CourseInfo with id %s not found.', $this->editEquipmentsCourseInfoCommand->getId()));
+        }
         try{
+            // Keep an original course equipments copy
+            $originalCourseResourceEquipments = $courseInfo->getCourseResourceEquipments();
+            // Fill course info with new values
+            $courseInfo = $this->editEquipmentsCourseInfoCommand->filledEntity($courseInfo);
             $courseInfo->setModificationDate(new \DateTime())
                 ->setLastUpdater($this->security->getUser());
             // Start transaction
             $this->courseInfoRepository->beginTransaction();
-
-            /** @var CourseResourceEquipment $courseResourceEquipment */
-            foreach ($courseInfo->getCourseResourceEquipments() as $offset => $courseResourceEquipment)
-            {
-                if (empty($courseResourceEquipment->getId()))
-                {
-                    $courseResourceEquipment->setId(Uuid::uuid4());
-                }
-            }
-
-            // Loop on original course resource equipments to detect ResourceEquipments must be removed
-            /** @var CourseResourceEquipment $courseResourceEquipment */
-            foreach ($originalCourseInfo as $courseResourceEquipment)
-            {
-                if (!$courseInfo->getCourseResourceEquipments()->contains($courseResourceEquipment))
-                {
+            // Loop on original course resource equipments to detect Resourceequipments must be removed
+            foreach ($originalCourseResourceEquipments as $courseResourceEquipment) {
+                if (!$courseInfo->getCourseResourceEquipments()->contains($courseResourceEquipment)) {
                     $this->courseResourceEquipmentRepository->delete($courseResourceEquipment);
                 }
             }
-
             // Update course infos
             $this->courseInfoRepository->update($courseInfo);
             $this->courseInfoRepository->commit();
