+++ src/AppBundle/Query/Course/EditPresentationCourseInfoQuery.php
@@ -2,18 +2,18 @@
 
 namespace AppBundle\Query\Course;
 
-use AppBundle\Entity\CourseInfo;
-use AppBundle\Entity\CourseTeacher;
+use AppBundle\Command\Course\EditPresentationCourseInfoCommand;
+use AppBundle\Exception\CourseInfoNotFoundException;
+use AppBundle\Query\QueryInterface;
 use AppBundle\Repository\CourseInfoRepositoryInterface;
 use AppBundle\Repository\CourseTeacherRepositoryInterface;
-use Ramsey\Uuid\Uuid;
 use Symfony\Component\Security\Core\Security;
 
 /**
  * Class EditPresentationCourseInfoQuery
  * @package AppBundle\Query\Course
  */
-class EditPresentationCourseInfoQuery
+class EditPresentationCourseInfoQuery implements QueryInterface
 {
 
     /**
@@ -25,6 +25,11 @@ class EditPresentationCourseInfoQuery
      * @var CourseTeacherRepositoryInterface
      */
     private $courseTeacherRepository;
+
+    /**
+     * @var EditPresentationCourseInfoCommand
+     */
+    private $editPresentationCourseInfoCommand;
 
     /**
      * @var Security
@@ -49,39 +54,44 @@ class EditPresentationCourseInfoQuery
     }
 
     /**
-     * @param CourseInfo $courseInfo
-     * @param CourseInfo $originalCourseInfo
+     * @param EditPresentationCourseInfoCommand $editPresentationCourseInfoCommand
+     * @return EditPresentationCourseInfoQuery
+     */
+    public function setEditPresentationCourseInfoCommand(EditPresentationCourseInfoCommand $editPresentationCourseInfoCommand): EditPresentationCourseInfoQuery
+    {
+        $this->editPresentationCourseInfoCommand = $editPresentationCourseInfoCommand;
+        return $this;
+    }
+
+    /**
+     * @throws CourseInfoNotFoundException
      * @throws \Exception
      */
-    public function execute(CourseInfo $courseInfo, CourseInfo $originalCourseInfo): void
+    public function execute(): void
     {
         try {
+            // Find CourseInfo
+            $courseInfo = $this->courseInfoRepository->find($this->editPresentationCourseInfoCommand->getId());
+        } catch(\Exception $e) {
+            throw $e;
+        }
+        if (is_null($courseInfo)) {
+            throw new CourseInfoNotFoundException(sprintf(
+                'CourseInfo with id %s not found.',
+                $this->editPresentationCourseInfoCommand->getId()
+            ));
+        }
+        try {
+            $originalCourseTeachers = $courseInfo->getCourseTeachers();
+            $courseInfo = $this->editPresentationCourseInfoCommand->filledEntity($courseInfo);
             $courseInfo->setModificationDate(new \DateTime())
                 ->setLastUpdater($this->security->getUser());
             $this->courseInfoRepository->beginTransaction();
-
-            dump($courseInfo);
-
-            /** @var CourseTeacher $courseTeacher */
-            foreach ($courseInfo->getCourseTeachers() as $offset => $courseTeacher)
-            {
-                if(empty($courseTeacher->getId()))
-                {
-                    $courseTeacher->setId(Uuid::uuid4());
-                }
-            }
-
-            dump($courseInfo);
-
-            /** @var CourseTeacher $courseTeacher */
-            foreach ($originalCourseInfo as $courseTeacher) {
+            foreach ($originalCourseTeachers as $courseTeacher) {
                 if (!$courseInfo->getCourseTeachers()->contains($courseTeacher)) {
                     $this->courseTeacherRepository->delete($courseTeacher);
                 }
             }
-
-            dump($courseInfo);
-
             $this->courseInfoRepository->update($courseInfo);
             $this->courseInfoRepository->commit();
         } catch(\Exception $e) {
